requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.110.0
h11==0.14.0
idna==3.10
pydantic==2.10.2
pydantic_core==2.27.1
python-dotenv==1.0.1
redis==5.0.3
requests==2.32.3
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.29.0
